Add Resultpage component
